pyproject.toml
[project]
name = "jwt-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyjwt>=2.10.1",
]
